+++ src/Querify.Nh.Tests/SubSpec/SubSpec.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using Xunit;
 using Xunit.Sdk;
-using System.Xml;
 using System.Diagnostics;
 using System.Reflection;
 using System.Linq;