+++ src/SolutionInfo.cs
@@ -7,7 +7,7 @@ using System.Reflection;
 
 [assembly: AssemblyDescription("Query and paging extensions for NHibernate")]
 [assembly: AssemblyProduct("Querify")]
-[assembly: AssemblyVersion("0.7.3.0")]
-[assembly: AssemblyFileVersion("0.7.3.0")]
-[assembly: AssemblyInformationalVersion("0.7.3.0")]
+[assembly: AssemblyVersion("0.8.0")]
+[assembly: AssemblyFileVersion("0.8.0")]
+[assembly: AssemblyInformationalVersion("0.8.0")]
 [assembly: AssemblyCopyright("Copyright (c) Lotpath 2014")]